Split APIHelper out into its own file
This allow to only import it when needed, speeding up the CLI for easy tasks,
especially the auto-completion provided by click [1]. This is most
noticable when site packages are enabled (as with the command "python -m
virtualenv --system-site-packages") or on global installations, like with
the official Debian package.

Before

	$time synadm > /dev/null
	real	0m0,610s
	user	0m0,537s
	sys	0m0,072s

After

	$time synadm > /dev/null
	real	0m0,086s
	user	0m0,077s
	sys	0m0,009s

[1]: https://click.palletsprojects.com/en/latest/shell-completion/
